ads updtes
